refactor(test): extract appwrite storage key utilities
Extract key encoding/decoding logic into reusable utilities
and simplify test configuration by centralizing project options
