+++ test/drivers/appwrite-storage.test.ts
@@ -1,7 +1,7 @@
 import { describe } from "vitest";
 import appwriteStorageDriver from "../../src/drivers/appwrite-storage.ts";
 import { testDriver } from "./utils.ts";
-import basex from "base-x";
+import { keyOptions } from "./appwrite.fixture.ts";
 
 const endpoint = process.env.VITE_APPWRITE_ENDPOINT;
 const projectId = process.env.VITE_APPWRITE_PROJECT_ID;
@@ -11,38 +11,20 @@ const apiKey = process.env.VITE_APPWRITE_API_KEY;
 describe.skipIf(!endpoint || !projectId || !bucketId)(
   "drivers: appwrite-storage",
   () => {
-    describe("keyStrategy: fileId", () => {
-      const base62 = basex(
-        "0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
-      );
-      const textEncoder = new TextEncoder();
-      const textDecoder = new TextDecoder();
-      const FILE_ID_SEPARATOR = "_" as const;
+    const projectOptions = {
+      endpoint: endpoint!,
+      projectId: projectId!,
+      bucketId: bucketId!,
+      apiKey,
+    };
 
+    describe("keyStrategy: fileId", () => {
       testDriver({
         driver: () =>
           appwriteStorageDriver({
             keyStrategy: "id",
-            endpoint: endpoint!,
-            projectId: projectId!,
-            bucketId: bucketId!,
-            apiKey: apiKey,
-            encodeKey(key, keySeparator) {
-              return key
-                .split(keySeparator)
-                .map((part) => {
-                  return base62.encode(textEncoder.encode(part));
-                })
-                .join(FILE_ID_SEPARATOR);
-            },
-            decodeKey(fileId, keySeparator) {
-              return fileId
-                .split(FILE_ID_SEPARATOR)
-                .map((part) => {
-                  return textDecoder.decode(base62.decode(part));
-                })
-                .join(keySeparator);
-            },
+            ...projectOptions,
+            ...keyOptions,
           }),
       });
     });
@@ -52,10 +34,8 @@ describe.skipIf(!endpoint || !projectId || !bucketId)(
         driver: () =>
           appwriteStorageDriver({
             keyStrategy: "name",
-            endpoint: endpoint!,
-            projectId: projectId!,
-            bucketId: bucketId!,
-            apiKey: apiKey,
+            ...projectOptions,
+            apiKey,
           }),
       });
     });
